Rename Deposit view
